revert config
